Remove optional parameters on updater and view functions
By marking parameters as optional, TypeScript requires verbose guards
when running in strict mode. I do not think these parameters should be
marked as optional, as Casium will always supply them internally.
However, it is up to the implementer to actually use them.

By marking them as non-optional, type checking and completion when
writing updater and view functions is a much more pleasant experience.

In order to achieve this, the `UpdaterDef` and `ContainerViewDef` type
aliases have been introduced, which mirror their existing counterparts
but do not contain optional parameters. These types are presented to
consumerrs of `container` when writing Updater and View functions.