Update 'master' to reflect currently published package (#46)
* Replace 'dev_tools' module with 'Instrumentation' API (#42)

* Replace 'dev_tools' module with 'Instrumenters' API

* Move State Manager reference to global scope and expose `cmdName`

- In order to allow the State Manager reference to persist across
  bundles, it needs to be stored in the global scope
- `cmdName`, although subject to change, is also useful for Dev Tools

* Upgrade TypeScript to 2.7

Allows use of constants as keys in type definitions

* Use Decorators to explicitly prefix Command Constructor names (#43)

* Remove optional parameters on updater and view functions (#36)

* Remove optional parameters on updater and view functions

By marking parameters as optional, TypeScript requires verbose guards
when running in strict mode. I do not think these parameters should be
marked as optional, as Casium will always supply them internally.
However, it is up to the implementer to actually use them.

By marking them as non-optional, type checking and completion when
writing updater and view functions is a much more pleasant experience.

In order to achieve this, the `UpdaterDef` and `ContainerViewDef` type
aliases have been introduced, which mirror their existing counterparts
but do not contain optional parameters. These types are presented to
consumerrs of `container` when writing Updater and View functions.

* Add support for function-returning Updaters to `UpdateResult`

Since an Updater can now return a function (see
e9712ee9722d3988f1ad2511c35e1c5bb4377aab), it follows that the
definition of `UpdateResult` should also include this as a permissible
type.

* Manually increment version number in package.json

Awkward merges led to this number being out-of-sync on `master`, so
manually incrementing it.